Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,10 +2,10 @@
 
 
 int main() {
-    int a, b, c;
-    scanf("%d", &a);
-    scanf("%d", &b);
-    scanf("%d", &c);
+    float a, b, c;
+    scanf("%f", &a);
+    scanf("%f", &b);
+    scanf("%f", &c);
     printf("Average: %.2f", (a+b+c)/3.0);
     return 0;
 }
